butano: some clangd warnings fixed

+++ butano/hw/include/bn_hw_dma.h
@@ -26,7 +26,7 @@ inline void copy_half_words(const void* source, int half_words, void* destinatio
     REG_DMA[3].cnt = 0;
     REG_DMA[3].src = source;
     REG_DMA[3].dst = destination;
-    REG_DMA[3].cnt = half_words | DMA_CPY16;
+    REG_DMA[3].cnt = unsigned(half_words) | DMA_CPY16;
 }
 
 inline void copy_words(const void* source, int words, void* destination)
@@ -34,7 +34,7 @@ inline void copy_words(const void* source, int words, void* destination)
     REG_DMA[3].cnt = 0;
     REG_DMA[3].src = source;
     REG_DMA[3].dst = destination;
-    REG_DMA[3].cnt = words | DMA_CPY32;
+    REG_DMA[3].cnt = unsigned(words) | DMA_CPY32;
 }
 
 inline void start_hdma(int channel, const uint16_t* source, int half_words, uint16_t* destination)
@@ -42,7 +42,7 @@ inline void start_hdma(int channel, const uint16_t* source, int half_words, uint
     REG_DMA[channel].cnt = 0;
     REG_DMA[channel].src = source;
     REG_DMA[channel].dst = destination;
-    REG_DMA[channel].cnt = half_words | DMA_HDMA;
+    REG_DMA[channel].cnt = unsigned(half_words) | DMA_HDMA;
 }
 
 inline void stop_hdma(int channel)
