+++ butano/src/bn_cameras_manager.cpp
@@ -39,7 +39,7 @@ namespace
     public:
         item_type items[max_items];
         alignas(int) uint8_t free_item_indexes_array[max_items];
-        int free_item_indexes_size = max_items;
+        uint16_t free_item_indexes_size = max_items;
         bool update = false;
     };
